fix initial-windows includes scratch buffer
Because of my display-buffer choices, a new buffer would be split
to show warnings. that buffer would also then be switched to the
home buffer, so I'd end up with a scratch buffer and a home buffer
and no warnings :sad:.

Fixed it by applying display-buffer customisations after init finished.

also:
* added evil args support and explored ruby alternatives.
* Enabled hindent-mode alongside haskell so evil open above and
  open below will work as expected, instead of leaving me to append
  another indent.
* made find-file-new-perspective use ivy completion instead of
  completing-read-file-name... now my ivy bindings work :smile:.
* also fix SPC s r doesn't use ripgrep