init.org (rename-buffer-file): allow moving to non-existant dirs
before it'd just crash saying you can't move to directory that doesn't
exist. Now the directory is silently made in the background when required.

also removed the restriction that you can't rename to a file that
emacs already has open (because of conflicting buffer names). that's
not really an issue seeing as emacs silently uses unique names for
buffers when appropriate.