init.org: add commands to toggle the display of headers
I've been exploring emacs in the terminal of late and seeing as I
generally have much less screen space, having a full line being taken
up by a header feels somewhat like a waste. Now a simple `SPC t h` and
the header will be hidden.

I've also modified my `mohkale-header-mode` renaming it to
`server-header-mode` and adding function to toggle the diplay of the
header; both globally and locally, such that you can prevent the
header appearing in all server buffers or it appearing in the current
buffer should emacsclient be invoked on it multiple times.