(init.org): Switch to editorconfig for indent control
Basically defers to the [[https://github.com/editorconfig/editorconfig-emacs][editorconfig]] package to assign tab-width for the current
buffer. Also added an override to `+indent-config' which uses the `editorconfig'
indent for the current file before my editors local config. You can override both
of these on a file-by-file basis by assigning the file local variable `+indent'.

This editorconfig a mandatory install for my emacs setup, but that's OK. Somewhere
down the line I might switch to using editorconfig for everything by default. There
seems to be a builtin-way to do this through a hook variable but I haven't looked
into it yet.