(lisp): Fix indentation
This has been annoying me for a while, turns out if you change
lisp-indent-offset lisp indentation in emacs will always use that
offset, even in special situations like `if` or `defun`.

editorconfig was aware of this and added an option to fix but it's
disabled by default :angry:.

This commit also re-adds a lisp-indent-function from doom. I'm not using
it yet, but it's there in case I ever need it.