(all-the-icons): Add nerd-font support
nerd-font was what all-the-icons was supposed to be. A single
font-package adding support for various other font-families except
nerd-font also patches the fonts so that there aren't any overlapping
character glyphs (a problem which terminal-users of all-the-icons are
bound to suffer from).

This patch adds in nerd-font as an adapter over all-the-icons, replacing
a lot of the replaceable parts of the all-the-icons alists with nerd-font
equivalents.