(config): Allow conditional tangling
A while back I added byte-compilation support to my emacs tangle script such that all
the lisp files produced by my config are subsequently byte-compiled to improve
performance. One downside to byte compilation is that any use-package or installation
expressions are evaluated immeadiately. That means if you have 50 packages, of which
25 are optional, all 50 will be installed even though only 25 will ever be used. This
both slows down compilation and startup times and on the whole leads to sluggish
performance.

This patch adapts my config to require my *+config.el* file while tangling. It then
uses those configuration options to decide which blocks to tangle and subsequently
results in the least amount of code needed to get a usable configuration.

Of course one down-side to all this is that my tangle-script no longer accurately
found the tangle dependencies of my config to tell whether it needs to be recompiled.
I've added a new script *./bin/tangle-targets* which performs a mock-tangle. It goes
through my config and all intermediate configs and outputs the files that would be
tangled to using a headless emacs process. This means tangling can now take upto 3
emacs processes but I believe the performance gains far outweigh the concerns.