(init): Fix tangle failure
For some weird dumb reason it looks like emacs won't make the lisp/
directory when tangling ~config.org~. This means tangling will crash
unless ~lisp/~ already exists :P. For now I've added a .gitignore to
that directory to force my emacs repo to have it available.

Another issue is that when I'm looking for tangle-targets (to detect
whether tangling needs to happen) I'm not actually tangling so loading
my config file to tell which blocks need to be tangled will immeadiately
crash. Before ~bin/tangle-targets~ used to tangle all but the first file
so this wasn't an issue but since now nothing is ever tangled it's gonna
cause my emacs to crash. For now I'm just evaluating the tangled config
code AND loading it, in the future I may have to clean this up.