(init): Fix tangle behaviour
This patch fixes a regression that was introduced back when I updated my
bin/tangle-targets script to not tangle any files ever. Part of my
previous tangle-setup had me loading my config file right at the start
of emacs but because tangle-targets doesn't tangle anything it can't
require my config and essentially just crashes unless a config file
exists.

The whole point of moving config to config.org was so that it would be
tangled by the time you evaluate the conditions in init.org, but seeing
as that's no longer possible I've opted to just move it back in. I've
also made my emacs evaluate the default config (while tangling) just in
case it can't require/load it on the first tangle-targets check.