(leader): Add use-package declaration
Now you can setup leader-key bindings directly from use-package.

This commit adds two new use-package keywords :leader and :lazy-leader.
The former assigns leader keys as soon as that use-package declaration
is evaluated. :lazy-leader does so after the package is loaded (when
:defer is not true).

You can specify a major mode, whether the mode is a minor mode or not
and a leader-prefix using plist props. The rest of the arguments are
treated as arguments to `spaceleader/define-key`.