(eglot): Add more flexible server start config
Before we either always started an eglot client or never did; now you
have the ability to connect to a server if its already running. The
default behaviour for now is to connect to a server if one is already
running in the current workspace.

For sh-mode I've set it up to always connect, for python-mode to only
connect when one exists across all active workspaces.