(company): Add set-company-backends! and cleanup bindings
Removed the ad-hoc hook based binding system with company-active-map.
Ended up avoiding it because of evil binding issues, so I added a
~"<override-state>" 'all~ call (didn't see that anywhere in the docs)
and everything seems to be working now. Also made the delay slightly
longer because it was too quick for my liking.

Lastly I've replaced ~company/configure-mode!~ with ~set-company-backend!~
which was taken from doom. There might be room for some improvement here
and it'll probably fail with some packages given how it's currently
setup (example: company-emoji), but otherwise everything should be fine.