(consult): Bifurcate from selectrum
consult is an excellent package but there's no direct relation between
it and selectrum. In fact consult can be plugged into multiple completion
backends so isolating my consult-config only to my selectrum-config is a
waste of a perfectly good and largely applicable package.
Now you can enable consult independent of selectrum, and therefore setup
consult with any other completion framework (that's consult/completing-read
compatible of course).
For now this can be done by including 'consult in ~+packages~, or
tangentially by installing selectrum.
As other completion frameworks are added I'll make it so that consult is
automatically enabled when it's compatible with those frameworks.