(consult): Add consult-line-or-ripgrep
I recently had to load a massive (50MB) JS data file and consult-line
was absolutely unusable. While I generally prefer built-ins to external
commands for the unique situation of a massive file it's more practical
to defer to an external process. Now C-s runs consult-line-or-ripgrep+
by default. For files under 2MB (or 1MB for org-mode (for some reason
that I partially inherited from counsel)) this'll just run consult-line.
For larger files it'll outsource grepping to ripgrep which should be
vastly more performant.