corfu: Add WIP support
Corfu is a new completion package for Emacs that works similar to
company but with a much lighter implementation that makes better use of
the builtin Emacs completion support. Currently I've just added corfu
and tried to set it up to be as similar to company as possible. You can
enable it as a new optional module.

Note: Corfu doesn't replace company. There are still several packages I
use that work only with company. To work around this I've set it up so
when corfu is enabled company is disabled and any company backends are
wrapped into compatible capf backends that corfu can present.