indent: Remove yank-indent-adjust
This is honestly more annoying than it is useful :/. I've started to
form a habit where I paste something, undo thinking it might've
re-indented when I didn't want it to, and then redo if it didn't.
Thinking about it knowing I want to indent after pasting is an explicit
choice. I think if I want to I can explicitly hit RET TAB or C-M--
afterwards, otherwise paste exactly what's on the clipboard for
consistency.