compilation: Fix race conditions with minibuffer and compilation auto-kill
Previously if you ran a compilation, a new compilation buffer was shown, and then you
quickly ran a command like find-file then your point would escape the minibuffer and
the compilation buffer would not be killed. The latter isn't solvable atm but the
former can and is fixed with this commit. We simply save the active window before
killing the compilation window.